backend/pyproject.toml
[tool.poetry]
name = "star_wars_characters"
version = "0.1.0"
description = ""
authors = ["Łukasz Skarżyński <skarzynski_lukasz@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

django = "^3.1"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
django-health-check = "^3.12.1"
djangorestframework = "^3.11.1"
petl = "^1.6.3"
psutil = "^5.7.2"
psycopg2 = "^2.8.5"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
Werkzeug = "^1.0.1"
coverage = "^5.2.1"
django-extensions = "^3.0.5"
django-querycount = "^0.7.0"
django-stubs = "^1.5.0"
django-test-migrations = "^1.0.0"
djangorestframework-stubs = "^1.2.0"
factory_boy = "^3.0.1"
flake8-django = "^1.1.1"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.3.0"
freezegun = "^0.3.15"
ipython = "^7.18.1"
nplusone = "^1.0.0"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^3.9.0"
pytest-factoryboy = "^2.0.3"
pytest-mock = "^3.3.1"
pytest-randomly = "^3.4.1"
pytest-timeout = "^1.4.2"
wemake-python-styleguide = "^0.14.1"

[build-system]
requires = ["poetry>=1.0.10"]
build-backend = "poetry.masonry.api"
